Add files via upload
The project aims to analyze the category "Pens, Pencils, and Writing Supplies" on Amazon using SQL to identify the most popular manufacturers. This will be achieved by examining the average number of reviews for products within this category. The analysis will provide insights into consumer preferences and help identify key players in the market.